pyproject.toml
[build-system]
requires = [
    "hatchling>=1.3.1",
]
build-backend = "hatchling.build"

[project]
name = "majik-obfuscate"
dynamic = ["version"]
description = "Data obfuscation library"
readme = "README.rst"
license = "Apache-2.0"
authors = [
    { name = "Shawn Michael", email = "python@monkeyspunk.net" },
]
keywords = [
    "obfuscation",
    "password",
    "storage",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Security",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]
dependencies = [
    "pycryptodomex",
]

[project.scripts]
obcat = "majik.obfuscate.obcat:start"
obwrite = "majik.obfuscate.obwrite:start"

[project.urls]
Homepage = "https://oss.majikindustries.com/python/majik-obfuscate/"

[tool.hatch.version]
path = "src/lib/majik/obfuscate/__init__.py"

[tool.hatch.build.targets.wheel]
packages = [
    "src/lib/majik",
]

[tool.hatch.build.targets.sdist]
include = [
    "/src/lib",
]

[tool.black]
line_length = 79
